requirements.txt
python-jobspy
google-genai
pandas
